ADDING new desing chagne and some improve
