+++ src/Hooks/useAdmin.jsx
@@ -1,7 +1,7 @@
 import { useQuery } from "@tanstack/react-query";
 import useAuth from "./useAuth";
-
-import axios from "axios";
+import useAxiosSecure from "./AxiosSequre";
+import useAxiosPublic from "./useAxiosPublic";
 
 const useAdmin = () => {
   const { user, loading } = useAuth();
@@ -11,7 +11,8 @@ const useAdmin = () => {
     enabled: !loading,
     queryFn: async () => {
       console.log("asking or checking is admin", user);
-      const res = await axios("http://localhost:5000/user/findAdmine");
+      // const res = await axios("http://localhost:5000/user/findAdmine");
+      const res = await useAxiosPublic.get(`/user/findAdmine/${user.email}`);
 
       return res.data?.admin;
     },
