+++ src/Routes/Routes.jsx
@@ -50,10 +50,11 @@ import PdfToPpt from "../Pages/PdfToPpt/PdfToPpt";
 import PdfToExcle from "../Pages/Pdftoexcel/Pdftoexcel";
 import OneUserReview from "../Component/UserReview/oneUserReview";
 import PptToPdf from "../Pages/PptToPdf/PptToPdf";
-import VoiceToText from "../Pages/VoicetoText/VoiceTospeech";
 
 import IssueFeedback from "../Pages/Dashboad/AdminePanelComponent/IssueFeedback/IssueFeed";
 import WordToPdf from "../Pages/Features/WordToPdf/WordToPdf";
+
+import SpeechToText from "../Pages/VoicetoText/SpeechToText";
 
 // import PdfToPpt from "../Pages/PdfToPpt/PdfToPpt";
 
@@ -98,8 +99,8 @@ export const router = createBrowserRouter([
       },
 
       {
-        path: "/speech",
-        element: <VoiceToText></VoiceToText>,
+        path: "/voice",
+        element: <SpeechToText></SpeechToText>,
       },
 
       {
